Clean code

+++ JsonMidiCreator/TalkieSPI/src/JsonTalker.cpp
@@ -22,7 +22,7 @@ bool JsonTalker::_is_led_on = false;
 
 
 
-bool JsonTalker::remoteSend(JsonObject json_message, bool as_reply, int target_index) {
+bool JsonTalker::remoteSend(JsonObject json_message, bool as_reply, uint8_t target_index) {
     if (_muted || _socket == nullptr) return false;
     json_message["f"] = _name;
     // 'c' = 0 means REMOTE communication (already set by socket's remoteSend)
